+++ app/layout.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from "next"
 import { Geist, Geist_Mono } from "next/font/google"
 import "./globals.css"
 import { Toaster } from "@/components/ui/sonner"
-import { ChatHistoryProvider } from "@/lib/chat-store/chat-history-provider"
+import { ChatHistoryProvider } from "@/lib/chat-store/chats/provider"
 import { APP_DESCRIPTION, APP_NAME } from "@/lib/config"
 import { ThemeProvider } from "next-themes"
 import Script from "next/script"
