+++ lib/chat-store/chats/api.ts
@@ -1,10 +1,66 @@
 import { readFromIndexedDB, writeToIndexedDB } from "@/lib/chat-store/persist"
 import type { Chat, ChatHistory } from "@/lib/chat-store/types"
 import { createClient } from "@/lib/supabase/client"
-import { MODEL_DEFAULT, SYSTEM_PROMPT_DEFAULT } from "../config"
-import { fetchClient } from "../fetch"
-import { API_ROUTE_CREATE_CHAT, API_ROUTE_UPDATE_CHAT_MODEL } from "../routes"
-import { getCachedChats } from "./history"
+import { MODEL_DEFAULT, SYSTEM_PROMPT_DEFAULT } from "../../config"
+import { fetchClient } from "../../fetch"
+import {
+  API_ROUTE_CREATE_CHAT,
+  API_ROUTE_UPDATE_CHAT_MODEL,
+} from "../../routes"
+
+export async function getCachedChats(): Promise<ChatHistory[]> {
+  const all = await readFromIndexedDB<ChatHistory>("chats")
+  return (all as ChatHistory[]).sort(
+    (a, b) => +new Date(b.created_at || "") - +new Date(a.created_at || "")
+  )
+}
+
+export async function fetchAndCacheChats(
+  userId: string
+): Promise<ChatHistory[]> {
+  const supabase = createClient()
+
+  const { data, error } = await supabase
+    .from("chats")
+    .select("id, title, created_at, model, system_prompt")
+    .eq("user_id", userId)
+    .order("created_at", { ascending: false })
+
+  if (!data || error) {
+    console.error("Failed to fetch chats:", error)
+    return []
+  }
+
+  await writeToIndexedDB("chats", data)
+  return data
+}
+
+export async function updateChatTitle(
+  id: string,
+  title: string
+): Promise<void> {
+  const supabase = createClient()
+  const { error } = await supabase.from("chats").update({ title }).eq("id", id)
+  if (error) throw error
+
+  const all = await getCachedChats()
+  const updated = (all as ChatHistory[]).map((c) =>
+    c.id === id ? { ...c, title } : c
+  )
+  await writeToIndexedDB("chats", updated)
+}
+
+export async function deleteChat(id: string): Promise<void> {
+  const supabase = createClient()
+  const { error } = await supabase.from("chats").delete().eq("id", id)
+  if (error) throw error
+
+  const all = await getCachedChats()
+  await writeToIndexedDB(
+    "chats",
+    (all as ChatHistory[]).filter((c) => c.id !== id)
+  )
+}
 
 export async function getChat(chatId: string): Promise<Chat | null> {
   const all = await readFromIndexedDB<Chat>("chats")
